refactor to allow storybook to re-use entry code
