+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "dogstack",
-  "version": "0.5.0-pre.5",
+  "version": "0.5.0-pre.10",
   "description": "a popular-choice grab-bag framework for teams working on production web apps",
   "main": "index.js",
   "browser": "browser.js",
